% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp_phase23.R
\name{run_resamp_inference_phase23}
\alias{run_resamp_inference_phase23}
\title{Function to perform resampling inference for Phase2/3 Seamless Trials}
\usage{
run_resamp_inference_phase23(
  mu0,
  mu1,
  mu2,
  sigma,
  n1,
  n2,
  alternative,
  alpha = 0.15,
  ws = seq(0, 1, length.out = 100),
  B1 = 100,
  B2 = 100,
  alpha_LCL = 0.025,
  htest_method = "mean"
)
}
\arguments{
\item{htest_method}{}
}
\description{
Function to perform resampling inference for Phase2/3 Seamless Trials
}
