% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_resamp_phase23_inference.R
\name{run_resamp_phase23_simulation}
\alias{run_resamp_phase23_simulation}
\title{Runs a simulation over n iterations for the resampling phase 2/3 inference}
\usage{
run_resamp_phase23_simulation(
  iter = 1000,
  mu0 = 0,
  mu1 = 0,
  mu2 = 0,
  n1 = 100,
  n2 = 100,
  ws = seq(0, 1, length.out = 100),
  alpha = 0.15,
  alpha_LCL = 0.025,
  htest_method = "mean",
  seed = 123
)
}
\arguments{
\item{seed}{}
}
\description{
Runs a simulation over n iterations for the resampling phase 2/3 inference
}
